service manager factories
