+++ src/main/java/org/spongepowered/mod/plugin/SpongeModPluginContainer.java
@@ -49,10 +49,12 @@ import net.minecraftforge.fml.common.versioning.VersionParser;
 import net.minecraftforge.fml.common.versioning.VersionRange;
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.plugin.PluginContainer;
+import org.spongepowered.api.service.ServiceManager;
 import org.spongepowered.common.SpongeImpl;
 import org.spongepowered.common.guice.SpongePluginGuiceModule;
 import org.spongepowered.common.plugin.AbstractPluginContainer;
 import org.spongepowered.common.plugin.PluginContainerExtension;
+import org.spongepowered.common.util.ServiceManagerFactory;
 
 import java.io.File;
 import java.net.URL;
@@ -65,6 +67,8 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+
+import javax.annotation.Nullable;
 
 // PluginContainer is implemented indirectly through the mixin to ModContainer
 public class SpongeModPluginContainer implements ModContainer, PluginContainerExtension {
@@ -85,6 +89,8 @@ public class SpongeModPluginContainer implements ModContainer, PluginContainerEx
     private LoadController controller;
 
     private Injector injector;
+
+    @Nullable private ServiceManager serviceManager;
 
     public SpongeModPluginContainer(String className, ModCandidate candidate, Map<String, Object> descriptor) {
         this.id = checkNotNull((String) descriptor.get("id"), "id");
@@ -394,6 +400,14 @@ public class SpongeModPluginContainer implements ModContainer, PluginContainerEx
     @Override
     public Injector getInjector() {
         return this.injector;
+    }
+
+    @SuppressWarnings("unused")
+    public ServiceManager getServiceManager() {
+        if (this.serviceManager == null) {
+            this.serviceManager = ServiceManagerFactory.createServiceManager();
+        }
+        return this.serviceManager;
     }
 
 }
